Replace `Package::git()` with `Project::git()`
